deployment-worker/requirements.txt
fastapi
uvicorn
ansible-runner
sse-starlette
psycopg2-binary
python-dotenv
pyyaml
pydantic
pika
hvac